+++ pkg/providers/iam/testdata/TestDataGroup/valid.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_iam_group" "test" {
+  group_id = 123
+}